Move URL parsing functions to their own module
This is an internal change only
